Fix 15 code review findings: overflow, crashes, races, dead deep links
Correctness:
- Fix Int*Int overflow in tip calculation (greeneats CheckoutViewModel.kt)
- Add OrderStatus.unknown + resilient decoder to greeneats iOS Models
- Fix PendingIntent requestCode=0 overwriting all notification taps (greeneats seller)
- Fix LocationManager unconditionally starting updates on foreground return
- Change OrderStatus.unknown to isActive=false (iOS + Android greeneats)
- Wire pendingRestaurantId deep link to HomeView NavigationStack
- Add LockedFormatter thread-safe wrapper to greeneats APIService
- Add writeTimeout/callTimeout to greeneats + greeneats-seller refreshClient
- Make silentResendOTP actually silent (seller AuthViewModel)
- Accept ZIP+4 format in address validation (koshereats consumer)
- Fix endTracking TOCTOU race in greeneats DeliveryActivityManager
- Remove leaked pushObserver in greeneats OrderTrackingViewModel.stop()
- Add past-time scheduledFor guard before placeOrder (iOS consumer)
- Show error on too-soon scheduled time instead of silent drop (Android)
- Parallelize N+1 cart sync with coroutineScope/async/awaitAll

+++ android/greeneats-consumer/app/src/main/java/com/greeneats/consumer/ui/viewmodels/CheckoutViewModel.kt
@@ -41,7 +41,7 @@ sealed interface TipChoice {
     fun label(subtotalCents: Int): String = when (this) {
         None -> "None"
         is Percent -> {
-            val cents = (subtotalCents * basisPoints) / 10_000
+            val cents = (subtotalCents.toLong() * basisPoints / 10_000).toInt()
             "${basisPoints / 100}%\n${cents.formatPrice()}"
         }
         Custom -> "Custom"
@@ -273,7 +273,7 @@ class CheckoutViewModel @Inject constructor(
         val subtotal = state.bundle?.subtotal ?: 0
         return when (val choice = state.tipChoice) {
             TipChoice.None -> 0
-            is TipChoice.Percent -> (subtotal * choice.basisPoints) / 10_000
+            is TipChoice.Percent -> (subtotal.toLong() * choice.basisPoints / 10_000).toInt()
             TipChoice.Custom -> {
                 val dollars = (state.customTipText.toDoubleOrNull() ?: 0.0).coerceAtLeast(0.0)
                 (dollars * 100).toInt().coerceAtMost(MAX_TIP_CENTS)
